chore: refactor some of the mods into more sensible directories
