+++ src/web/mod.rs
@@ -3,8 +3,8 @@ pub mod dockerflow;
 pub mod extractors;
 pub mod handlers;
 pub mod middleware;
+pub mod adm;
 #[cfg(test)]
 mod test;
-mod user_agent;
 
 pub use dockerflow::DOCKER_FLOW_ENDPOINTS;
